uart/pl011: Compute rclk when requested by the boot loader.
If the boot loader sets xo=-1, then we can compute the rclk to use from
the baud rate and divisors used. The practical effect of this change is
that we'll be able to change the baud-rate of the console. But it will
also report it which may be helpful for other serial ports in the same
system.

Sponsored by:		Netflix