uart/ns8250: Compute rclk when the boot loader tells us too.
When the boot loader sets xo=-1, it's telling us that the passed-in
baudrate has been programmed into the hardware is correct. This allows
us to recover the rclk so login works (since previously we'd set
the divisor with a bogus default rclk). Compute the rclk after
reading the divisor and doing the inverse math we use to set the
baud rate.

Sponsored by:		Netflix