kboot: Add support for UEFI booting amd64
Now that we can find the EFI SYSTEM TABLE via the Linux boot_param
"zero-page", add code to the trampole to snag it and write it to our
metedta area.

The physical address of the system table is needed for UEFI to do
anything, but it's not published at all in the Linux sysfs. The Linux
pandora boot loader copies it, and a few other things, from the
currently running Linux boot params. This is why Linux can work without
it being observable by kexec-tools. However, this is not an OS agnsotic
way to publish this critical information. To cope, create a minimal
boot_param struct and use that to document where things are so we can
use sensible #defines for accessing the two halves of the value in the
boot_param. Put the address where to write this on the top of the stack
so the trampoline code can pop it off easily.

Sponsored by:		Netflix