loader: Remove unneeded includes
We've not needed to include link.h in the loader for a long time. I
shoud have removed it in f38658e1408c rather than going link.h ->
sys/sys_link.h. It's purely a userland shared library thing. Remove it
now, since it's not needed. Also remove a few headers that are redundant
with stand.h in this environment.

Sponsored by:		Netflix