umass: Move illegal opcode return to a function
Move illegal opcode completion of ccb to its own function. da can cope
with a number of different failures now (it didn't used to), so we
should return the unnsupported commands as illegal rather than faking
completion. This will allow da to, in the future, do more intelligent
things in response to these quirks. Other SIMs will need to do this
too.