umass: Document the ideal that it's better to fail bad commands
Generally, lying to the periph drivers about commands working is a bad
idea. They cannot get a global insight into the actual capacities of
the device when commands just work. There's good reason to intercept
bad commands at times, and to prevent the device from being exposed to
commands that, for example, hand the device. However, upper layer
periph drivers is the better place to cope with devices that don't
support the commands since it can make other inferences about the
device when it has wider knowledge.

But this represents a change in philosphy, so document that as
well. This shift happened for some commands a while ago. Many of the
quirks existed only to avoid whining errors on boot that were benign
and were a big burden. So da and cd slowly stopped the whining. Except
in cases where the command hangs the drive, we should limit new quirks
and avoid overly broad quirks (things work, but are less performant or
reliable, again mostly to avoid harmless errors that are no longer
printed).

Other sims will likely need some help with this.