umass: Bring in small fix from NetBSD's umass
When completing a request for UFI, don't fail the request on
non-zero asc/ascq values if we've done a request sense.

This idea is from umass.c 1.100 by mycroft. He used it to help
elminate the INQUIRY_SHORT quirk that we still have. However, it will
make little difference because we treat both return values the same
for CBI. And it appears we have (maybe bogusly) some devices that
specify this quirk that aren't CBI.

Sponsored by:		Netflix