mpi3mr V3 : Add mpi3mr Driver to sys/conf/files and GENERIC file for Kernel Inclusion
Registered the mpi3mr driver source files in sys/conf/files, enabling
it to be compiled into the base kernel image. This matches the approach
used for existing Broadcom drivers such as mrsas, mps, and mpr.

Also updated sys/conf/NOTES with an entry for the mpi3mr driver.

With this change, the mpi3mr driver will be built as part of the base
kernel, allowing automatic loading during boot when supported hardware
is detected.

Below changes are suggested by Chuck.

Changes from V1 to V2:

- Corrected the file name typo mistake in /sys/conf/files from mpi3mr_app.c.c to mpi3mr_app.c
- Added mpi3mr driver entries in "sys/amd64/conf/GENERIC" and "sys/i386/conf/GENERIC"

Reviewed by:	ssaxena, chs, imp
Differential Revision:	https://reviews.freebsd.org/D49754